pytorch_poc/pyproject.toml
[project]
name = "pytorch-poc"
version = "0.1.0"
description = "Pytorch VAE proof of concept"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
